Adding Leases to parachain end-point.

+++ src/main/polkauction/core/model/Lease.kt
@@ -3,4 +3,4 @@ package polkauction.core.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Lease(val leaseIndexPeriod: String, val account: String, val deposit: Int)
+data class Lease(val leaseIndexPeriod: String, val account: String, val deposit: Double)
